Rails 7 prework part3: Restructure Evidence engine
