+++ services/QuillLMS/engines/evidence/lib/evidence/check/opinion.rb
