+++ services/QuillLMS/engines/evidence/lib/evidence/check/regex_sentence.rb
