+++ services/QuillLMS/engines/evidence/lib/evidence/html_tag_remover.rb
