+++ services/QuillLMS/engines/evidence/lib/evidence/open_ai.rb
