+++ services/QuillLMS/engines/evidence/lib/evidence/open_ai/completion.rb
@@ -3,7 +3,7 @@
 module Evidence
   module OpenAI
     class Completion < Evidence::ApplicationService
-      include Evidence::OpenAI::Concerns::API
+      include Evidence::OpenAI::Concerns::Api
       include Evidence::OpenAI::Concerns::SentenceResults
 
       ENDPOINT = '/completions'
