+++ services/QuillLMS/engines/evidence/lib/evidence/open_ai/concerns/sentence_results.rb
