+++ services/QuillLMS/engines/evidence/lib/evidence/plagiarism_check.rb
