+++ services/QuillLMS/engines/evidence/lib/evidence/profanity.rb
