+++ services/QuillLMS/engines/evidence/lib/evidence/safe_first_lowercaser.rb
