+++ services/QuillLMS/engines/evidence/lib/evidence/spelling_check.rb
