+++ services/QuillLMS/engines/evidence/lib/evidence/string_normalizer.rb
