+++ services/QuillLMS/engines/evidence/lib/evidence/synthetic/generators/spelling.rb
