+++ services/QuillLMS/engines/evidence/lib/evidence/synthetic/generators/translation.rb
