+++ services/QuillLMS/engines/evidence/lib/evidence/synthetic/labeled_result.rb
