+++ services/QuillLMS/engines/evidence/lib/evidence/synthetic/manual_types.rb
