+++ services/QuillLMS/engines/evidence/lib/evidence/version.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 module Evidence
-  VERSION = "0.0.6"
+  class Version
+    VERSION = "0.0.6"
+  end
 end
