+++ services/QuillLMS/engines/evidence/lib/generators/quill_model/templates/model.rb.tt
