+++ services/QuillLMS/engines/evidence/lib/generators/quill_scaffold/quill_scaffold_generator.rb
@@ -1,11 +1,14 @@
 require 'rails/generators/rails/resource/resource_generator'
 
-class QuillScaffoldGenerator < Rails::Generators::ResourceGenerator
-  remove_hook_for :resource_controller
-  remove_class_option :actions
+module Generators
+  module QuillScaffold
+    class QuillScaffoldGenerator < Rails::Generators::ResourceGenerator
+      remove_hook_for :resource_controller
+      remove_class_option :actions
 
-  class_option :resource_route, type: :boolean
+      class_option :resource_route, type: :boolean
 
-  hook_for :scaffold_controller, required: true
-
+      hook_for :scaffold_controller, required: true
+    end
+  end
 end
