+++ services/QuillLMS/engines/evidence/lib/generators/quill_scaffold_controller/templates/quill_api_controller.rb.tt
