+++ services/QuillLMS/engines/evidence/spec/dummy/config/routes.rb
@@ -1,4 +1,3 @@
 Rails.application.routes.draw do
-
   mount Evidence::Engine => "/evidence", as: :evidence
 end
